transaction signer library + py03 module compilation
