+++ transaction-signer/Cargo.toml
@@ -9,6 +9,11 @@ name = "transaction-signer"
 path = "src/bin/main.rs"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[lib]
+name = "transaction_signer_lib"
+path = "src/lib.rs"
+crate-type = ["cdylib"]
+
 [dependencies]
 mc-account-keys = { path = "../mobilecoin/account-keys" }
 mc-common = { path = "../mobilecoin/common", default-features = false, features = ["loggers"] }
@@ -29,3 +34,6 @@ serde = { version = "1.0", default-features = false, features = ["alloc", "deriv
 serde_json = { version = "1.0", features = ["preserve_order"] }
 structopt = "0.3"
 tiny-bip39 = "1.0"
+pyo3 = { version = "0.17.3", features = ["extension-module"] }
+
+
